make use of the synchronous setter methods to eliminate controller.awaitCompletion() inconsistencies
The problem was that actions scheduled for async execution with the async() function (like setValue()) would themselves cause scheduling of async actions FROM the async context (this would introduce a new action in the cue after the await-action), and was due to the listeners set up in the controllers constructor. So awaitCompletion() would have to be called multiple times to get the desired effect. With the new sync setters this issue can be avoided. The design sucks tho, it's hard to tell how data flows and what influences what.