refactor Segment.isOn() function to look up state in Game.activatedEdges and Game.terminals
This hopefully fixes the nasty bug where LED-anims completely break if the game is muted (first of all ???, second of all: the Segment.isOn() function was used to check if an anim should be aborted. And because this is done asynchronously, the boolean returned by this function was sometimes not updated in time. But this would only matter if the delay from sound instantiation wasn't present.)